requirements.txt

streamlit
faiss-cpu
numpy
requests
